requirements.txt
numpy==1.26.0
matplotlib==3.8.0
scikit-learn==1.3.2
Pillow==11.0.0
gradio
